// framework/glsp-client/tasklist-glsp/src/framework/common/types/types.ts
export type ValidationResult = {
    /** URI containing `id` and `type` */
    uri: string;
    /**
     * Represents multiple diagnostics, such as a compiler errors or warnings.
     * @see {@link Diagnostic}
     */
    diagnostics: Diagnostic[];
};
// ...
export type GlspServices = {
    BatchValidationHandler: BatchValidationHandler;
    DocumentAstHandler: DocumentAstHandler;
    ScopingInformationListener: ScopingInformationListener;
    BatchValidationListener: BatchValidationListener;
    GlspConnection: GlspConnection;
    ExternalNodeInformationService: ExternalNodeInformationService;
};

/** Additional shared Langium services */
export type GlspLangiumSharedGlspServices = {
    glsp: GlspServices;
};

/** Langium shared services extended with {@link GlspServices} */
export type GlspLangiumSharedServices = LangiumSharedServices & GlspLangiumSharedGlspServices;